refactor(kwin-effect): extract window-lifecycle slots to plasmazoneseffect/window_lifecycle.cpp
Pulls 8 window lifecycle methods out of plasmazoneseffect.cpp:
slotWindowAdded, slotWindowClosed, slotWindowActivated,
setupWindowConnections, notifyWindowClosed, notifyWindowActivated,
findWindowById, findAllWindowsById. Drops the main file from 4391 to
3792 lines.
